+++ CHANGES.md
@@ -1,6 +1,12 @@
 TS-Interval Changes
 ===================
 
+1.0.2
+-----
+
+- Add `IntervalData` interface, add `Interval.data()` and constructor support.
+- Export function to format Span as string (`formatSpan`)
+
 1.0.1
 -----